+++ Trust/Foundation/PushableCoordinator.swift
@@ -1,0 +1,10 @@
+// Copyright SIX DAY LLC. All rights reserved.
+
+import Foundation
+
+typealias RootCoordinator = Coordinator & RootViewControllerProvider
+
+protocol PushableCoordinator {
+    func pushCoordinator(_ coordinator: RootCoordinator)
+    func popCoordinator(_ coordinator: RootCoordinator)
+}
