+++ Trust/Foundation/RootViewControllerProvider.swift
@@ -1,0 +1,8 @@
+// Copyright SIX DAY LLC. All rights reserved.
+
+import Foundation
+import UIKit
+
+protocol RootViewControllerProvider: class {
+    var rootViewController: UIViewController { get }
+}
